works with debug but not release

+++ build.sh
@@ -31,11 +31,11 @@ build_package() {
 }
 
 if [ ! -f ${INSTALL_DIR}/lib/zlib.lib ] ; then
-build_package zlib-1.2.11 
-if [ ! -f ${INSTALL_DIR}/lib/zlib.lib ] ; then
-    echo "zlib build failed"
-    exit 1
-fi
+    build_package zlib-1.2.11 
+    if [ ! -f ${INSTALL_DIR}/lib/zlib.lib ] ; then
+	echo "zlib build failed"
+	exit 1
+    fi
 else
     echo zlib already installed
 fi
@@ -121,7 +121,7 @@ else
 fi
 
 if [ ! -f ${INSTALL_DIR}/lib/gdal_i.lib ] ; then
-cmd "/c gdal.bat"
+    cmd "/c gdal.bat"
 if [ ! -f ${INSTALL_DIR}/lib/gdal_i.lib ] ; then
     echo gdal build failed
     exit 1
@@ -129,48 +129,57 @@ fi
 else
     echo gdal already installed
 fi
-# ------ simage -----
-if [ ! -f ${INSTALL_DIR}/lib/simage1.lib ] ; then
-if [ -d simage-1.7.0/build/msvc12 ] ; then
-    rm -rf simage-1.7.0/build/msvc12
-fi
-tar xaf simage_vs2013.tar
-cmd "/c simage.bat"
-if [ ! -f ${INSTALL_DIR}/lib/simage1.lib ] ; then
-    echo simage build failed
-    exit 1
-fi
-else
-    echo simage already installed
-fi
 
 
-# ------ coin -----
-if [ ! -f ${INSTALL_DIR}/lib/coin3.lib ] ; then
-if [ ! -f Coin-3.1.3/build/msvc9/include/Inventor/system/inttypes.h.orig ] ; then
-    echo patching Coin
-    patch -b Coin-3.1.3/build/msvc9/include/Inventor/system/inttypes.h < coin.patch
-    patch -b Coin-3.1.3/build/misc/build-docs.bat < coin_build-docs.patch
-    tar xaf coin_buildUpgrade.tbz
-fi
-cmd "/c coin.bat"
-if [ ! -f ${INSTALL_DIR}/lib/coin3.lib ] ; then
-    echo coin3 build failed
-    exit 1
-fi
-else
-    echo coin already installed
+if [ X${BIN_COIN} == X ] ; then
+    # ------ simage -----
+    if [ ! -f ${INSTALL_DIR}/lib/simage1.lib ] ; then
+	if [ -d simage-1.7.0/build/msvc12 ] ; then
+	    rm -rf simage-1.7.0/build/msvc12
+	fi
+	tar xaf simage_vs2013.tar
+	cmd "/c simage.bat"
+	if [ ! -f ${INSTALL_DIR}/lib/simage1.lib ] ; then
+	    echo simage build failed
+	    exit 1
+	fi
+    else
+	echo simage already installed
+    fi
+
+
+    # ------ coin -----
+    if [ ! -f ${INSTALL_DIR}/lib/coin3.lib ] ; then
+	if [ ! -f Coin-3.1.3/build/msvc9/include/Inventor/system/inttypes.h.orig ] ; then
+	    echo patching Coin
+	    patch -b Coin-3.1.3/build/msvc9/include/Inventor/system/inttypes.h < coin.patch
+	    patch -b Coin-3.1.3/build/misc/build-docs.bat < coin_build-docs.patch
+	    tar xaf coin_buildUpgrade.tbz
+	fi
+	cmd "/c coin.bat"
+	if [ ! -f ${INSTALL_DIR}/lib/coin3.lib ] ; then
+	    echo coin3 build failed
+	    exit 1
+	fi
+	if [ ! -f ${INSTALL_DIR}/lib/coin3d.lib ] ; then
+	    echo coin3 debug build failed
+	    exit 1
+	fi
+    else
+	echo coin already installed
+    fi
 fi
 
-echo ---------------- $1 ---------------- 
-if [ ! -f ${INSTALL_DIR}/lib/osg.lib ] ; then
-
+echo ---------------- OSG ---------------- 
 if [ ! -d OpenSceneGraph-3.4.0/cmake_build ] ; then 
     echo mkdir OpenSceneGraph-3.4.0/cmake_build
     mkdir OpenSceneGraph-3.4.0/cmake_build
 fi
 pushd OpenSceneGraph-3.4.0/cmake_build
 
+
+if [ X${BIN_COIN} == X ]  ; then
+    echo build against compiled COIN
 cmake -G "Visual Studio 12 2013 Win64" \
       -DCMAKE_INSTALL_PREFIX=${INSTALL_DIR} \
       \
@@ -189,8 +198,8 @@ cmake -G "Visual Studio 12 2013 Win64" \
       -DGDAL_LIBRARY=${INSTALL_DIR}/lib/gdal_i.lib \
       \
       -DINVENTOR_INCLUDE_DIR=${INSTALL_DIR}/include \
-      -DINVENTOR_LIBRARY_DEBUG=${INSTALL_DIR}/lib/coin3d.lib \
       -DINVENTOR_LIBRARY_RELEASE=${INSTALL_DIR}/lib/coin3.lib \
+      -DINVENTOR_LIBRARY_DEBUG=${INSTALL_DIR}/lib/coin3d.lib \
       \
       -DJPEG_INCLUDE_DIR=${INSTALL_DIR}/include \
       -DJPEG_LIBRARY=${INSTALL_DIR}/lib/libjpeg.lib \
@@ -199,6 +208,39 @@ cmake -G "Visual Studio 12 2013 Win64" \
       -DPNG_LIBRARY_DEBUG=${INSTALL_DIR}/lib/libpng16d.lib \
       -DPNG_LIBRARY_RELEASE=${INSTALL_DIR}/lib/libpng16.lib \
       ..
+else
+    echo build against bin dist COIN
+cmake -G "Visual Studio 12 2013 Win64" \
+      -DCMAKE_INSTALL_PREFIX=${INSTALL_DIR} \
+      \
+      -DZLIB_INCLUDE_DIR=${INSTALL_DIR}/include \
+      -DZLIB_LIBRARY_DEBUG=${INSTALL_DIR}/lib/zlibd.lib \
+      -DZLIB_LIBRARY_RELEASE=${INSTALL_DIR}/lib/zlib.lib \
+      \
+      -DTIFF_INCLUDE_DIR=${INSTALL_DIR}/include \
+      -DTIFF_LIBRARY_DEBUG=${INSTALL_DIR}/lib/tiffd.lib \
+      -DTIFF_LIBRARY_RELEASE=${INSTALL_DIR}/lib/tiff.lib \
+      \
+      -DCURL_INCLUDE_DIR=${INSTALL_DIR}/include \
+      -DCURL_LIBRARY=${INSTALL_DIR}/lib/libcurl_imp.lib \
+      \
+      -DGDAL_INCLUDE_DIR=${INSTALL_DIR}/include \
+      -DGDAL_LIBRARY=${INSTALL_DIR}/lib/gdal_i.lib \
+      \
+      -DINVENTOR_INCLUDE_DIR=${INSTALL_DIR}/binCoin/include \
+      -DINVENTOR_LIBRARY_DEBUG=${INSTALL_DIR}/binCoin/lib/coin3d.lib \
+      -DINVENTOR_LIBRARY_RELEASE=${INSTALL_DIR}/binCoin/lib/coin3.lib \
+      \
+      -DJPEG_INCLUDE_DIR=${INSTALL_DIR}/include \
+      -DJPEG_LIBRARY=${INSTALL_DIR}/lib/libjpeg.lib \
+      \
+      -DPNG_PNG_INCLUDE_DIR=${INSTALL_DIR}/include \
+      -DPNG_LIBRARY_DEBUG=${INSTALL_DIR}/lib/libpng16d.lib \
+      -DPNG_LIBRARY_RELEASE=${INSTALL_DIR}/lib/libpng16.lib \
+      ..
+fi
+
+
 
 # For some reason we get bogus libraries inserted into the project file
 for f in `find . -name \*.vcxproj -print` ; do
@@ -207,14 +249,16 @@ for f in `find . -name \*.vcxproj -print` ; do
     mv foo $f
 done
 
-
-cmake --build . --config Release > build_r.log
-cmake --build . --config Release --target install > install_r.log
+echo building debug version
 cmake --build . --config Debug > build_d.log
+echo installing debug version
 cmake --build . --config Debug --target install> install_d.log
+echo building release version
+cmake --build . --config Release > build_r.log
+echo installing release version
+cmake --build . --config Release --target install > install_r.log
 
 popd
-
 
 if [ ! -f ${INSTALL_DIR}/lib/osg.lib ] ; then
     echo osg build failed
@@ -224,6 +268,4 @@ fi
 if [ ! -f ${INSTALL_DIR}/lib/osgd.lib ] ; then
     echo osg debug build failed
     exit 1
-fi
-    
 fi
